Finally fixing and refactoring the tests
