+++ tests/routes/explorer/results-panel/test.ts
@@ -1,0 +1,153 @@
+import { expect, type Route } from '@playwright/test';
+import { test } from '../../../custom-context';
+import {
+  conceptsDetailPath,
+  detailResponseCat,
+  searchResults as mockData,
+  searchResultPath,
+} from '../../../mock-data';
+
+test.describe('Results Panel', () => {
+  test('Result panel bar and button shows', async ({ page }) => {
+    // Given
+    await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+    await page.goto('/explorer?search=somedata');
+
+    // Then
+    await expect(page.locator('#side-panel-bar')).toBeVisible();
+    await expect(page.locator('#results-panel-toggle')).toBeVisible();
+    await page.locator('#results-panel-toggle').click();
+  });
+  test('Result toggle button opens and closes the results panel', async ({ page }) => {
+    // Given
+    await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+    await page.goto('/explorer?search=somedata');
+
+    //When
+    await page.locator('#results-panel-toggle').click();
+
+    // Then
+    await expect(page.locator('#results-panel')).toBeVisible();
+
+    //When
+    await page.locator('#results-panel-toggle').click();
+
+    // Then
+    await expect(page.locator('#results-panel')).not.toBeVisible();
+  });
+  test('Result panel shows 0 results on an error', async ({ page }) => {
+    // Given
+    await page.goto('/explorer?search=somedata');
+    await page.locator('#results-panel-toggle').click();
+
+    // Then
+    await expect(page.locator('#result-count')).toBeVisible();
+    await expect(page.locator('#result-count')).toHaveText('0');
+  });
+  test('Result panel shows the correct number of results', async ({ page }) => {
+    // Given
+    await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+    await page.route('*/**/picsure/query/sync', async (route: Route) =>
+      route.fulfill({ body: '9999' }),
+    );
+    await page.goto('/explorer?search=somedata');
+    await page.locator('#results-panel-toggle').click();
+
+    // Then
+    await expect(page.locator('#result-count')).toBeVisible();
+    await expect(page.locator('#result-count')).not.toHaveText('0');
+  });
+  test('Result panel shows no filters added when there are no filters', async ({ page }) => {
+    // Given
+    await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+    await page.route('*/**/picsure/query/sync', async (route: Route) =>
+      route.fulfill({ body: '9999' }),
+    );
+    await page.goto('/explorer?search=somedata');
+    await page.locator('#results-panel-toggle').click();
+
+    // Then
+    await expect(page.getByText('No filters added')).toBeVisible();
+  });
+  test('Export button hidden when no filters or exports are added', async ({ page }) => {
+    // Given
+    await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+    await page.route('*/**/picsure/query/sync', async (route: Route) =>
+      route.fulfill({ body: '9999' }),
+    );
+    await page.goto('/explorer?search=somedata');
+
+    // When
+    await page.locator('#results-panel-toggle').click();
+
+    // Then
+    await expect(page.getByText('No filters added')).toBeVisible();
+    await expect(page.locator('#export-data-button')).not.toBeVisible();
+  });
+  test('Export button hidden when count is 0', async ({ page }) => {
+    // Given
+    await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+    await page.route('*/**/picsure/query/sync', async (route: Route) =>
+      route.fulfill({ body: '0' }),
+    );
+    await page.goto('/explorer?search=somedata');
+
+    // When
+    await page.locator('#results-panel-toggle').click();
+
+    // Then
+    await expect(page.locator('#results-panel')).toBeVisible();
+    await expect(page.locator('#export-data-button')).not.toBeVisible();
+  });
+  test('Export button hidden when count is not 0 and there is no filters', async ({ page }) => {
+    // Given
+    await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+    await page.route('*/**/picsure/query/sync', async (route: Route) =>
+      route.fulfill({ body: '9999' }),
+    );
+
+    // When
+    await page.goto('/explorer?search=somedata');
+    await page.locator('#results-panel-toggle').click();
+
+    // Then
+    await expect(page.locator('#results-panel')).toBeVisible();
+    await expect(page.locator('#export-data-button')).not.toBeVisible();
+  });
+  test('Clear All clears exports and filters', async ({ page }) => {
+    // Given
+    await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+    await page.route(`${conceptsDetailPath}${detailResponseCat.dataset}`, async (route: Route) =>
+      route.fulfill({ json: detailResponseCat }),
+    );
+    await page.route('*/**/picsure/query/sync', async (route: Route) =>
+      route.fulfill({ body: '9999' }),
+    );
+    await page.goto('/explorer?search=age');
+
+    const expectedRowIds = mockData.content.map((row) => row.name);
+    const tableBody = page.locator('tbody');
+    await expect(tableBody).toBeVisible();
+
+    const firstRow = tableBody.locator('tr').nth(0);
+    const filterIcon = firstRow.locator('td').last().locator('button').nth(1);
+    await filterIcon.click();
+    const firstFilter = await getOption(page);
+    await firstFilter.click();
+    const addFilterButton = page.getByTestId('add-filter');
+    await addFilterButton.click();
+    await expect(page.getByTestId(`added-filter-${expectedRowIds[0]}`)).toBeVisible();
+
+    const secondRow = tableBody.locator('tr').nth(1);
+    const exportButton = secondRow.locator('td').last().locator('button').last();
+    await exportButton.click();
+    await expect(page.getByTestId(`added-export-${expectedRowIds[1]}`)).toBeVisible();
+
+    // When
+    await page.getByTestId('clear-all-results-btn').click();
+    await page.getByTestId('modal').getByRole('button', { name: 'Yes' }).click();
+
+    // Then
+    await expect(page.getByText('No filters added')).toBeVisible();
+  });
+});
