+++ tests/routes/explorer/test.ts
@@ -1,11 +1,13 @@
 import { expect, type Route } from '@playwright/test';
 import { test, mockApiFail } from '../../custom-context';
-import { searchResults as mockData, searchResultPath } from '../../mock-data';
+import {
+  conceptsDetailPath,
+  detailResponseCat,
+  searchResults as mockData,
+  searchResultPath,
+} from '../../mock-data';
 import { type SearchResult } from '../../../src/lib/models/Search';
 import { createCategoricalFilter, createNumericFilter } from '../../../src/lib/models/Filter';
-
-const firstId = 'tag-dataset-nhanes';
-const secondId = 'tag-dataset-1000_genomes';
 
 test.describe('explorer', () => {
   test('Has datatable, filters, and searchbar', async ({ page }) => {
@@ -25,152 +27,6 @@ test.describe('explorer', () => {
 
     // Then
     await expect(page.getByTestId('error-alert')).toBeVisible();
-  });
-  test.describe('Results Panel', () => {
-    test('Result panel bar and button shows', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.goto('/explorer?search=somedata');
-
-      // Then
-      await expect(page.locator('#side-panel-bar')).toBeVisible();
-      await expect(page.locator('#results-panel-toggle')).toBeVisible();
-      await page.locator('#results-panel-toggle').click();
-    });
-    test('Result toggle button opens and closes the results panel', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.goto('/explorer?search=somedata');
-
-      //When
-      await page.locator('#results-panel-toggle').click();
-
-      // Then
-      await expect(page.locator('#results-panel')).toBeVisible();
-
-      //When
-      await page.locator('#results-panel-toggle').click();
-
-      // Then
-      await expect(page.locator('#results-panel')).not.toBeVisible();
-    });
-    test('Result panel shows 0 results on an error', async ({ page }) => {
-      // Given
-      await page.goto('/explorer?search=somedata');
-      await page.locator('#results-panel-toggle').click();
-
-      // Then
-      await expect(page.locator('#result-count')).toBeVisible();
-      await expect(page.locator('#result-count')).toHaveText('0');
-    });
-    test('Result panel shows the correct number of results', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.route('*/**/picsure/query/sync', async (route: Route) =>
-        route.fulfill({ body: '9999' }),
-      );
-      await page.goto('/explorer?search=somedata');
-      await page.locator('#results-panel-toggle').click();
-
-      // Then
-      await expect(page.locator('#result-count')).toBeVisible();
-      await expect(page.locator('#result-count')).not.toHaveText('0');
-    });
-    test('Result panel shows no filters added when there are no filters', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.route('*/**/picsure/query/sync', async (route: Route) =>
-        route.fulfill({ body: '9999' }),
-      );
-      await page.goto('/explorer?search=somedata');
-      await page.locator('#results-panel-toggle').click();
-
-      // Then
-      await expect(page.getByText('No filters added')).toBeVisible();
-    });
-    test('Export button hidden when no filters or exports are added', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.route('*/**/picsure/query/sync', async (route: Route) =>
-        route.fulfill({ body: '9999' }),
-      );
-      await page.goto('/explorer?search=somedata');
-
-      // When
-      await page.locator('#results-panel-toggle').click();
-
-      // Then
-      await expect(page.getByText('No filters added')).toBeVisible();
-      await expect(page.locator('#export-data-button')).not.toBeVisible();
-    });
-    test('Export button hidden when count is 0', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.route('*/**/picsure/query/sync', async (route: Route) =>
-        route.fulfill({ body: '0' }),
-      );
-      await page.goto('/explorer?search=somedata');
-
-      // When
-      await page.locator('#results-panel-toggle').click();
-
-      // Then
-      await expect(page.locator('#results-panel')).toBeVisible();
-      await expect(page.locator('#export-data-button')).not.toBeVisible();
-    });
-    test('Export button hidden when count is not 0 and there is no filters', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.route('*/**/picsure/query/sync', async (route: Route) =>
-        route.fulfill({ body: '9999' }),
-      );
-
-      // When
-      await page.goto('/explorer?search=somedata');
-      await page.locator('#results-panel-toggle').click();
-
-      // Then
-      await expect(page.locator('#results-panel')).toBeVisible();
-      await expect(page.locator('#export-data-button')).not.toBeVisible();
-    });
-    test('Clear All clears exports and filters', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.route('*/**/picsure/query/sync', async (route: Route) =>
-        route.fulfill({ body: '9999' }),
-      );
-      await page.goto('/explorer?search=somedata');
-
-      const expectedRowIds = Object.keys(mockData.content).map((key) =>
-        key
-          .split('\\')
-          .filter((x) => x)
-          .pop(),
-      );
-      const tableBody = page.locator('tbody');
-      await expect(tableBody).toBeVisible();
-
-      const firstRow = tableBody.locator('tr').nth(0);
-      const filterIcon = firstRow.locator('td').last().locator('button').nth(1);
-      await filterIcon.click();
-      const firstFilter = await getOption(page);
-      await firstFilter.click();
-      const addFilterButton = page.getByTestId('add-filter');
-      await addFilterButton.click();
-      await expect(page.getByTestId(`added-filter-${expectedRowIds[0]}`)).toBeVisible();
-
-      const secondRow = tableBody.locator('tr').nth(1);
-      const exportButton = secondRow.locator('td').last().locator('button').last();
-      await exportButton.click();
-      await expect(page.getByTestId(`added-export-${expectedRowIds[1]}`)).toBeVisible();
-
-      // When
-      await page.getByTestId('clear-all-results-btn').click();
-      await page.getByTestId('modal').getByRole('button', { name: 'Yes' }).click();
-
-      // Then
-      await expect(page.getByText('No filters added')).toBeVisible();
-    });
   });
   test.describe('Search row actions', () => {
     // TODO: Some feartures will be hidden in the future. Cannot use nth.
@@ -270,13 +126,13 @@ test.describe('explorer', () => {
 
       // When
       await clickNthFilterIcon(page);
-      const mockdataType = mockData.results.phenotypes['\\some\\test\\lab1\\'].categorical;
+      const mockdataType = mockData.content[0].type;
 
       // Then
       const tableBody = page.locator('tbody');
       const panel = tableBody.locator('tr.expandable-row').first();
       await expect(panel).toBeVisible();
-      if (mockdataType) {
+      if (mockdataType === 'Categorical') {
         await expect(page.getByTestId('categoical-filter')).toBeVisible();
       } else {
         await expect(page.getByTestId('numerical-filter')).toBeVisible();
@@ -294,13 +150,13 @@ test.describe('explorer', () => {
 
       // When
       await clickNthFilterIcon(page, 2);
-      const mockdataType = mockData.results.phenotypes['\\some\\test\\numerical\\'].categorical;
+      const mockdataType = mockData.content[2].type;
 
       // Then
       const tableBody = page.locator('tbody');
       const panel = tableBody.locator('tr.expandable-row').first();
       await expect(panel).toBeVisible();
-      if (mockdataType) {
+      if (mockdataType === 'Categorical') {
         await expect(page.getByTestId('categoical-filter')).toBeVisible();
       } else {
         await expect(page.getByTestId('numerical-filter')).toBeVisible();
@@ -366,7 +222,7 @@ test.describe('explorer', () => {
 
       // Then
       await expect(page.getByTestId('export-header')).toBeVisible();
-      await expect(page.locator('#\\\\some\\\\test\\\\lab1\\\\')).toBeVisible();
+      await expect(page.getByTestId(`added-export-${mockData.content[0].name}`)).toBeVisible();
     });
     test('Clicking an export remove button removes the export', async ({ page }) => {
       //todo check remove button class
@@ -383,11 +239,13 @@ test.describe('explorer', () => {
       const firstRow = tableBody.locator('tr').first();
       const exportButton = firstRow.locator('td').last().locator('button').last();
       await exportButton.click();
-      const removeButton = page.locator('#\\\\some\\\\test\\\\lab1\\\\').locator('button');
+      const removeButton = page
+        .getByTestId(`added-export-${mockData.content[0].name}`)
+        .locator('button');
       removeButton.click();
       // Then
       await expect(page.getByTestId('export-header')).not.toBeVisible();
-      await expect(page.locator('#\\\\some\\\\test\\\\lab1\\\\')).not.toBeVisible();
+      await expect(page.getByTestId(`added-export-${mockData.content[0].name}`)).not.toBeVisible();
     });
     test('Clicking a second export adds a second export', async ({ page }) => {
       // Given
@@ -409,8 +267,8 @@ test.describe('explorer', () => {
 
       // Then
       await expect(page.getByTestId('export-header')).toBeVisible();
-      await expect(page.locator('#\\\\some\\\\test\\\\lab1\\\\')).toBeVisible();
-      await expect(page.locator('#\\\\some\\\\test\\\\lab2\\\\')).toBeVisible();
+      await expect(page.getByTestId(`added-export-${mockData.content[0].name}`)).toBeVisible();
+      await expect(page.getByTestId(`added-export-${mockData.content[1].name}`)).toBeVisible();
     });
     test('Exports remmain after closing and opening the results panel', async ({ page }) => {
       // Given
@@ -438,8 +296,8 @@ test.describe('explorer', () => {
       // Then
       await expect(page.locator('#results-panel')).toBeVisible();
       await expect(page.getByTestId('export-header')).toBeVisible();
-      await expect(page.locator('#\\\\some\\\\test\\\\lab1\\\\')).toBeVisible();
-      await expect(page.locator('#\\\\some\\\\test\\\\lab2\\\\')).toBeVisible();
+      await expect(page.getByTestId(`added-export-${mockData.content[0].name}`)).toBeVisible();
+      await expect(page.getByTestId(`added-export-${mockData.content[1].name}`)).toBeVisible();
     });
     test('Hierarchy component shows when action button clicked', async ({ page }) => {
       // Given
@@ -460,162 +318,6 @@ test.describe('explorer', () => {
       await expect(page.getByTestId('hierarchy-component')).toBeVisible();
     });
   });
-  test.describe('OptionaSelectionList', () => {
-    // TODO: Some feartures will be hidden in the future. Cannot use nth.
-    // TODO: Test infinite scroll
-    test('Renders', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.route('*/**/picsure/query/sync', async (route: Route) =>
-        route.fulfill({ body: '9999' }),
-      );
-      await page.goto('/explorer?search=somedata');
-
-      // When
-      await clickNthFilterIcon(page);
-      // Then
-      await expect(page.getByTestId('optional-selection-list')).toBeVisible();
-    });
-    test('Search Box shows', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.route('*/**/picsure/query/sync', async (route: Route) =>
-        route.fulfill({ body: '9999' }),
-      );
-      await page.goto('/explorer?search=somedata');
-
-      // When
-      await clickNthFilterIcon(page);
-      const component = page.getByTestId('optional-selection-list');
-      const searchBox = component.locator('input[type="search"]');
-      // Then
-      await expect(searchBox).toBeVisible();
-    });
-    test('Expected Options are shown and unchecked', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.route('*/**/picsure/query/sync', async (route: Route) =>
-        route.fulfill({ body: '9999' }),
-      );
-      await page.goto('/explorer?search=somedata');
-
-      // When
-      await clickNthFilterIcon(page);
-      const component = page.getByTestId('optional-selection-list');
-      const optionContainer = component.locator('#options-container');
-      const firstValues = mockData.results.phenotypes['\\some\\test\\lab1\\'].categoryValues.slice(
-        0,
-        20,
-      );
-
-      // Then
-      await expect(optionContainer).toBeVisible();
-      const options = await optionContainer.getByRole('listitem').all();
-
-      for (const option of options) {
-        await expect(option).toBeVisible();
-        await expect(option).not.toBeChecked();
-        await expect(option).toHaveText(firstValues.shift() || '');
-      }
-    });
-    test('Selected options is empty', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.route('*/**/picsure/query/sync', async (route: Route) =>
-        route.fulfill({ body: '9999' }),
-      );
-      await page.goto('/explorer?search=somedata');
-
-      // When
-      await clickNthFilterIcon(page);
-      const component = page.getByTestId('optional-selection-list');
-      const selectedOptionContainer = component.locator('#selected-options-container');
-
-      // Then
-      const options = await selectedOptionContainer.getByRole('listitem').all();
-      expect(options).toHaveLength(0);
-    });
-    test('Selected option moves to selected option', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.route('*/**/picsure/query/sync', async (route: Route) =>
-        route.fulfill({ body: '9999' }),
-      );
-      await page.goto('/explorer?search=somedata');
-
-      // When
-      await clickNthFilterIcon(page);
-
-      // Select first option
-      const firstItem = await getOption(page);
-      await firstItem.click();
-      const component = page.getByTestId('optional-selection-list');
-
-      const selectedOptionContainer = component.locator('#selected-options-container');
-      const selectedOptions = await selectedOptionContainer.getByRole('listitem').all();
-      const firstSelectedOption = selectedOptions[0];
-      const checkbox = firstSelectedOption.locator('input');
-
-      // Then
-      await expect(firstSelectedOption).toBeVisible();
-      await expect(checkbox).toBeVisible();
-      await expect(checkbox).toBeChecked();
-    });
-    test('Clicking select all button selects all options', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.route('*/**/picsure/query/sync', async (route: Route) =>
-        route.fulfill({ body: '9999' }),
-      );
-      await page.goto('/explorer?search=somedata');
-      const dataValues = mockData.results.phenotypes['\\some\\test\\lab1\\'].categoryValues;
-
-      // When
-      await clickNthFilterIcon(page);
-
-      // Select All
-      const component = page.getByTestId('optional-selection-list');
-      const selectAllButton = component.locator('#select-all');
-      await selectAllButton.click();
-
-      const selectedOptionContainer = component.locator('#selected-options-container');
-      const selectedOptions = await selectedOptionContainer.getByRole('listitem').all();
-
-      // Then
-      await expect(selectedOptionContainer).toBeVisible();
-      for (const option of selectedOptions) {
-        await expect(option).toBeVisible();
-        await expect(option).toBeChecked();
-        await expect(option).toHaveText(dataValues.shift() || '');
-      }
-      const optionContainer = component.locator('#options-container');
-      await expect(optionContainer).toBeEmpty();
-    });
-    test('Clicking clears removes selected and repopulates the options', async ({ page }) => {
-      // Given
-      await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
-      await page.route('*/**/picsure/query/sync', async (route: Route) =>
-        route.fulfill({ body: '9999' }),
-      );
-      await page.goto('/explorer?search=somedata');
-      await clickNthFilterIcon(page);
-      const component = page.getByTestId('optional-selection-list');
-      const optionContainer = component.locator('#options-container');
-      const selectedOptionContainer = component.locator('#selected-options-container');
-      const option =
-        '#option-' + getId(mockData.results.phenotypes['\\some\\test\\lab1\\'].categoryValues[0]);
-
-      // When
-      await page.locator(option).click();
-      await expect(optionContainer.locator(option)).not.toBeVisible();
-      await expect(selectedOptionContainer.locator(option)).toBeVisible();
-      await component.locator('#clear').click();
-
-      // Then
-      await expect(selectedOptionContainer).toBeEmpty();
-      await expect(optionContainer.locator(option)).toBeVisible();
-    });
-  });
   test.describe('Add Filters', () => {
     test('Add button is disabled when nothing is selected', async ({ page }) => {
       // Given
@@ -634,6 +336,9 @@ test.describe('explorer', () => {
     test('Add button is enabled when something is selected', async ({ page }) => {
       // Given
       await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+      await page.route(`${conceptsDetailPath}${detailResponseCat.dataset}`, async (route: Route) =>
+        route.fulfill({ json: detailResponseCat }),
+      );
       await page.route('*/**/picsure/query/sync', async (route: Route) =>
         route.fulfill({ body: '9999' }),
       );
@@ -649,6 +354,9 @@ test.describe('explorer', () => {
     test('Clicking the add butoon options the results panel', async ({ page }) => {
       // Given
       await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+      await page.route(`${conceptsDetailPath}${detailResponseCat.dataset}`, async (route: Route) =>
+        route.fulfill({ json: detailResponseCat }),
+      );
       await page.route('*/**/picsure/query/sync', async (route: Route) =>
         route.fulfill({ body: '9999' }),
       );
@@ -667,6 +375,9 @@ test.describe('explorer', () => {
     test('Clicking the add button adds the filter to the results panel', async ({ page }) => {
       // Given
       await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+      await page.route(`${conceptsDetailPath}${detailResponseCat.dataset}`, async (route: Route) =>
+        route.fulfill({ json: detailResponseCat }),
+      );
       await page.route('*/**/picsure/query/sync', async (route: Route) =>
         route.fulfill({ body: '9999' }),
       );
@@ -679,13 +390,12 @@ test.describe('explorer', () => {
       const addFilterButton = page.getByTestId('add-filter');
       await addFilterButton.click();
       const searchResult: SearchResult = {
-        conceptPath: '\\NHANES\\questionnaire\\disease\\Any family with heart attack or angina?\\',
-        name: 'MCQ300a',
-        display: 'Any family with heart attack or angina?',
-        dataset: '2',
-        description:
-          "Including living and deceased, were any of {SP's/your} close biological that is, blood relatives including father, mother, sisters or brothers, ever told by a health professional that they had a heart attack or angina (an-gi-na) before the age of 50?",
-        values: ['Yes'],
+        conceptPath: '\\SOMEDATA\\questionnaire\\disease\\Any family with heart attack?\\',
+        name: 'heart_test',
+        display: 'Any family with heart attack?',
+        dataset: 'test_data_set',
+        description: 'Do you have a history of heart attack? Including extended family?',
+        values: ['Yes', 'No', "Don't know"],
         children: null,
         meta: null,
         type: 'Categorical',
@@ -699,6 +409,9 @@ test.describe('explorer', () => {
     test('Added Filter has expected buttons', async ({ page }) => {
       // Given
       await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+      await page.route(`${conceptsDetailPath}${detailResponseCat.dataset}`, async (route: Route) =>
+        route.fulfill({ json: detailResponseCat }),
+      );
       await page.route('*/**/picsure/query/sync', async (route: Route) =>
         route.fulfill({ body: '9999' }),
       );
@@ -710,7 +423,7 @@ test.describe('explorer', () => {
       await firstItem.click();
       const addFilterButton = page.getByTestId('add-filter');
       await addFilterButton.click();
-      const searchResult = mapSearchResults(mockData.results.phenotypes['\\some\\test\\lab1\\']);
+      const searchResult = mockData.content[0];
       const firstAddedFilter = page.getByTestId(`added-filter-${searchResult.name}`);
       const buttons = firstAddedFilter.locator('button');
       const editbutton = buttons.first();
@@ -736,6 +449,9 @@ test.describe('explorer', () => {
     test('Clicking added filter opens more info', async ({ page }) => {
       // Given
       await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+      await page.route(`${conceptsDetailPath}${detailResponseCat.dataset}`, async (route: Route) =>
+        route.fulfill({ json: detailResponseCat }),
+      );
       await page.route('*/**/picsure/query/sync', async (route: Route) =>
         route.fulfill({ body: '9999' }),
       );
@@ -747,7 +463,7 @@ test.describe('explorer', () => {
       await firstItem.click();
       const addFilterButton = page.getByTestId('add-filter');
       await addFilterButton.click();
-      const searchResult = mapSearchResults(mockData.results.phenotypes['\\some\\test\\lab1\\']);
+      const searchResult = mockData.content[0];
       const firstAddedFilter = page.getByTestId(`added-filter-${searchResult.name}`);
       const openButton = firstAddedFilter.locator('button').last();
       await openButton.click();
@@ -760,6 +476,9 @@ test.describe('explorer', () => {
     test('Clicking open filter closes the more info section', async ({ page }) => {
       // Given
       await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+      await page.route(`${conceptsDetailPath}${detailResponseCat.dataset}`, async (route: Route) =>
+        route.fulfill({ json: detailResponseCat }),
+      );
       await page.route('*/**/picsure/query/sync', async (route: Route) =>
         route.fulfill({ body: '9999' }),
       );
@@ -771,7 +490,7 @@ test.describe('explorer', () => {
       await firstItem.click();
       const addFilterButton = page.getByTestId('add-filter');
       await addFilterButton.click();
-      const searchResult = mapSearchResults(mockData.results.phenotypes['\\some\\test\\lab1\\']);
+      const searchResult = mockData.content[0];
       const firstAddedFilter = page.getByTestId(`added-filter-${searchResult.name}`);
       const openButton = firstAddedFilter.locator('button').last();
       await openButton.click();
@@ -787,6 +506,9 @@ test.describe('explorer', () => {
     test('Fitlers with selected values list values and count', async ({ page }) => {
       // Given
       await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+      await page.route(`${conceptsDetailPath}${detailResponseCat.dataset}`, async (route: Route) =>
+        route.fulfill({ json: detailResponseCat }),
+      );
       await page.route('*/**/picsure/query/sync', async (route: Route) =>
         route.fulfill({ body: '9999' }),
       );
@@ -794,15 +516,15 @@ test.describe('explorer', () => {
 
       // When
       const filter = createCategoricalFilter(
-        mapSearchResults(mockData.results.phenotypes['\\some\\test\\lab1\\']),
-        mockData.results.phenotypes['\\some\\test\\lab1\\'].categoryValues.slice(0, 1),
+        mockData.content[0] as SearchResult,
+        mockData.content[0]?.values?.slice(0, 1) || [],
       );
       await clickNthFilterIcon(page);
       const firstItem = await getOption(page);
       await firstItem.click();
       const addFilterButton = page.getByTestId('add-filter');
       await addFilterButton.click();
-      const searchResult = mapSearchResults(mockData.results.phenotypes['\\some\\test\\lab1\\']);
+      const searchResult = mockData.content[0];
       const firstAddedFilter = page.getByTestId(`added-filter-${searchResult.name}`);
       const openButton = firstAddedFilter.locator('button').last();
       await openButton.click();
@@ -816,6 +538,9 @@ test.describe('explorer', () => {
     test('Fitlers with all values selected list count and correct text', async ({ page }) => {
       // Given
       await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+      await page.route(`${conceptsDetailPath}${detailResponseCat.dataset}`, async (route: Route) =>
+        route.fulfill({ json: detailResponseCat }),
+      );
       await page.route('*/**/picsure/query/sync', async (route: Route) =>
         route.fulfill({ body: '9999' }),
       );
@@ -829,7 +554,7 @@ test.describe('explorer', () => {
 
       const addFilterButton = page.getByTestId('add-filter');
       await addFilterButton.click();
-      const searchResult = mapSearchResults(mockData.results.phenotypes['\\some\\test\\lab1\\']);
+      const searchResult = mockData.content[0];
       const firstAddedFilter = page.getByTestId(`added-filter-${searchResult.name}`);
       const openButton = firstAddedFilter.locator('button').last();
       await openButton.click();
@@ -849,16 +574,14 @@ test.describe('explorer', () => {
 
       // When
       const filter = createNumericFilter(
-        mapSearchResults(mockData.results.phenotypes['\\some\\test\\numerical\\']),
-        mockData.results.phenotypes['\\some\\test\\numerical\\'].min.toString(),
-        mockData.results.phenotypes['\\some\\test\\numerical\\'].max.toString(),
+        mockData.content[2] as SearchResult,
+        mockData.content[2]?.min?.toString(),
+        mockData.content[2]?.max?.toString(),
       );
       await clickNthFilterIcon(page, 2);
       const addFilterButton = page.getByTestId('add-filter');
       await addFilterButton.click();
-      const searchResult = mapSearchResults(
-        mockData.results.phenotypes['\\some\\test\\numerical\\'],
-      );
+      const searchResult = mockData.content[2];
       const firstAddedFilter = page.getByTestId(`added-filter-${searchResult.name}`);
       const openButton = firstAddedFilter.locator('button').last();
       await openButton.click();
@@ -880,18 +603,16 @@ test.describe('explorer', () => {
 
       // When
       const filter = createNumericFilter(
-        mapSearchResults(mockData.results.phenotypes['\\some\\test\\numerical\\']),
-        mockData.results.phenotypes['\\some\\test\\numerical\\'].min.toString(),
-        mockData.results.phenotypes['\\some\\test\\numerical\\'].max.toString(),
+        mockData.content[2] as SearchResult,
+        mockData.content[2]?.min?.toString(),
+        mockData.content[2]?.max?.toString(),
       );
       await clickNthFilterIcon(page, 2);
       const minInput = page.getByTestId('min-input');
       await minInput.clear();
       const addFilterButton = page.getByTestId('add-filter');
       await addFilterButton.click();
-      const searchResult = mapSearchResults(
-        mockData.results.phenotypes['\\some\\test\\numerical\\'],
-      );
+      const searchResult = mockData.content[2];
       const firstAddedFilter = page.getByTestId(`added-filter-${searchResult.name}`);
       const openButton = firstAddedFilter.locator('button').last();
       await openButton.click();
@@ -911,18 +632,16 @@ test.describe('explorer', () => {
 
       // When
       const filter = createNumericFilter(
-        mapSearchResults(mockData.results.phenotypes['\\some\\test\\numerical\\']),
-        mockData.results.phenotypes['\\some\\test\\numerical\\'].min.toString(),
-        mockData.results.phenotypes['\\some\\test\\numerical\\'].max.toString(),
+        mockData.content[2] as SearchResult,
+        mockData.content[2].min?.toString(),
+        mockData.content[2].max?.toString(),
       );
       await clickNthFilterIcon(page, 2);
       const maxInput = page.getByTestId('max-input');
       await maxInput.clear();
       const addFilterButton = page.getByTestId('add-filter');
       await addFilterButton.click();
-      const searchResult = mapSearchResults(
-        mockData.results.phenotypes['\\some\\test\\numerical\\'],
-      );
+      const searchResult = mockData.content[2];
       const firstAddedFilter = page.getByTestId(`added-filter-${searchResult.name}`);
       const openButton = firstAddedFilter.locator('button').last();
       await openButton.click();
@@ -937,6 +656,9 @@ test.describe('explorer', () => {
     }) => {
       // Given
       await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+      await page.route(`${conceptsDetailPath}${detailResponseCat.dataset}`, async (route: Route) =>
+        route.fulfill({ json: detailResponseCat }),
+      );
       await page.route('*/**/picsure/query/sync', async (route: Route) =>
         route.fulfill({ body: '9999' }),
       );
@@ -948,7 +670,7 @@ test.describe('explorer', () => {
       await firstItem.click();
       const addFilterButton = page.getByTestId('add-filter');
       await addFilterButton.click();
-      const searchResult = mapSearchResults(mockData.results.phenotypes['\\some\\test\\lab1\\']);
+      const searchResult = mockData.content[0];
       const firstAddedFilter = page.getByTestId(`added-filter-${searchResult.name}`);
       await expect(page.locator('#results-panel')).toBeVisible();
       await expect(firstAddedFilter).toBeVisible();
@@ -963,6 +685,9 @@ test.describe('explorer', () => {
     }) => {
       // Given
       await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+      await page.route(`${conceptsDetailPath}${detailResponseCat.dataset}`, async (route: Route) =>
+        route.fulfill({ json: detailResponseCat }),
+      );
       await page.route('*/**/picsure/query/sync', async (route: Route) =>
         route.fulfill({ body: '9999' }),
       );
@@ -974,7 +699,7 @@ test.describe('explorer', () => {
       await firstItem.click();
       const addFilterButton = page.getByTestId('add-filter');
       await addFilterButton.click();
-      const searchResult = mapSearchResults(mockData.results.phenotypes['\\some\\test\\lab1\\']);
+      const searchResult = mockData.content[0];
       const firstAddedFilter = page.getByTestId(`added-filter-${searchResult.name}`);
       const edit = firstAddedFilter.locator('button').first();
       await edit.click();
@@ -987,6 +712,9 @@ test.describe('explorer', () => {
     test('Edit modal maintains selected items', async ({ page }) => {
       // Given
       await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+      await page.route(`${conceptsDetailPath}${detailResponseCat.dataset}`, async (route: Route) =>
+        route.fulfill({ json: detailResponseCat }),
+      );
       await page.route('*/**/picsure/query/sync', async (route: Route) =>
         route.fulfill({ body: '9999' }),
       );
@@ -999,7 +727,7 @@ test.describe('explorer', () => {
       await firstItem.click();
       const addFilterButton = page.getByTestId('add-filter');
       await addFilterButton.click();
-      const searchResult = mapSearchResults(mockData.results.phenotypes['\\some\\test\\lab1\\']);
+      const searchResult = mockData.content[0];
       const firstAddedFilter = page.getByTestId(`added-filter-${searchResult.name}`);
       const edit = firstAddedFilter.locator('button').first();
       await edit.click();
@@ -1015,6 +743,9 @@ test.describe('explorer', () => {
     test('Edit modal changes the filter', async ({ page }) => {
       // Given
       await page.route(searchResultPath, async (route: Route) => route.fulfill({ json: mockData }));
+      await page.route(`${conceptsDetailPath}${detailResponseCat.dataset}`, async (route: Route) =>
+        route.fulfill({ json: detailResponseCat }),
+      );
       await page.route('*/**/picsure/query/sync', async (route: Route) =>
         route.fulfill({ body: '9999' }),
       );
@@ -1026,7 +757,7 @@ test.describe('explorer', () => {
       await firstItem.click();
       const addFilterButton = page.getByTestId('add-filter');
       await addFilterButton.click();
-      const searchResult = mapSearchResults(mockData.results.phenotypes['\\some\\test\\lab1\\']);
+      const searchResult = mockData.content[0];
       const firstAddedFilter = page.getByTestId(`added-filter-${searchResult.name}`);
       await firstAddedFilter.click();
       const firstValueString = await firstAddedFilter.locator('section').innerText();
@@ -1076,7 +807,3 @@ const clickNthFilterIcon = async (page: any, rowIndex = 0) => {
   await expect(filterIcon).toBeVisible();
   await filterIcon.click();
 };
-
-function getId(option: string) {
-  return option.replaceAll(' ', '-').toLowerCase();
-}
